Trick the generate command
For some reason the checker does not accept the path – I get the error "Cannot find module '~/content/about.md'". If I just move it to a variable the generate command works, but the dev server no longer loads the file. This ugly hack makes both happy, and me sad.