Fix the sidebar layout for 1024px wide windows
E.g. half screen MBP. In this configuration the sidebar was a bit too far to the right edge, but this makes the margins on both sides balance better. The space between the words in the logo on the front page had to be a bit narrower to make it look good, though.