Redesign the way the article list is included
The sidebar slot has been removed and the article list can now just be enabled or disabled with a prop.

The way the article list is shown on the front page has been redesigned to an inline style. I never liked the way the list rendered next to the title/logo.