Add a hamburger menu top left when the sidebar is not shown
It's too hard to find the article list otherwise.

This was a major CSS pain to get working, there was all sorts of alignment problems, and small differences between Chrome and Mobile Safari. The main problem is that neither the menu toggle, the article list or the logo have a width, which makes it hard to get the logo to float to the right, and the article list to align well with the menu toggle. I managed to get it working through using visibility instead of display, hiding the article list and setting its height to 1px makes it affect the flex layout, but not take up actual space. Using relative positioning also made it possible to align the article list with the menu toggle.