Override the default theme to get code styling to work
Instead of a completely custom theme, this overrides the default theme. I couldn't figure out any other way to get the code block styling to work.

The rest of the changes are attempts to get the styling to look more or less like what it does in the Nuxt version.