Make /articles/xyz/ work again
In the old site, /articles/xyz/ was the canonical URL, but in the new only /articles/xyz (without slash at the end) works. I haven't been able to figure out how to make a redirect for it. All search engines have the URLs with a slash at the end, and traffic has plummeted since the new site was launched.

This new build script makes a copy of every article in a way that should work. I don't know if this is a viable long term solution.